requirements.txt
asgiref==3.2.10
Django==3.0.8
django-redis==4.12.1
djangorestframework==3.11.0
pytz==2020.1
redis==3.5.3
sqlparse==0.3.1
